requirements/requirements.txt
absl-py
arviz>=0.11.2
immutabledict>=2.0.0
jax>=0.3.18
jaxlib>=0.3.18
matplotlib==3.7.2
numpy>=1.23
numpyro>=0.9.2
pandas>=1.1.5
scipy
seaborn==0.12.2
scikit-learn
statsmodels>=0.13.0
tensorflow>=2.7.2
